Fix User Error Null

+++ final-project-fe/Dtos/Users/User.cs
@@ -6,7 +6,7 @@
 		public string Email { get; set; }
 		public string Phone { get; set; }
 		public string Password { get; set; }
-		public int Point { get; set; }
+		public decimal Point { get; set; }
 		public bool IsBanned { get; set; } = false;
 		public DateTime CreateAt { get; set; } 
 		public DateTime? UpdateAt { get; set; }
